pyproject.toml
[project]
name = "yukky-sandbox-guide-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.66.1",
]

[tool.uv]
dev-dependencies = [
    "grpcio-testing>=1.66.1",
    "grpcio-tools>=1.66.1",
    "python-dotenv>=1.0.1",
]
